Moved ui model to the presentation layer.
