+++ lib/presentation/mapper/movie_mapper.dart
@@ -1,6 +1,6 @@
 import 'package:star_wars_flutter/domain/model/movie.dart';
 import 'package:star_wars_flutter/presentation/mapper/mapper.dart';
-import 'package:star_wars_flutter/ui/model/movie_view_model.dart';
+import 'package:star_wars_flutter/presentation/model/movie_view_model.dart';
 
 class MovieMapper implements Mapper<MovieViewModel, Movie> {
   @override
