--- lib/presentation/model/character_view.dart
@@ -1,22 +1,0 @@
-class CharacterView {
-  CharacterView(
-      {required this.id,
-      required this.birthDay,
-      required this.eyeColor,
-      required this.gender,
-      required this.hairColor,
-      required this.height,
-      required this.homeWorld,
-      required this.mass,
-      required this.name});
-
-  final String id;
-  final String birthDay;
-  final String eyeColor;
-  final String gender;
-  final String hairColor;
-  final String height;
-  final String homeWorld;
-  final String mass;
-  final String name;
-}
