+++ lib/presentation/model/character_view_model.dart
