--- lib/presentation/model/movie_view.dart
@@ -1,23 +1,0 @@
-/// Representation for a [MovieView] instance for this layers Model representation
-class MovieView {
-  MovieView(
-      {required this.id,
-      required this.title,
-      required this.director,
-      required this.releaseDate,
-      required this.producer,
-      required this.plot,
-      required this.url,
-      required this.characters,
-      this.imdbRating});
-
-  final int id;
-  final String title;
-  final String director;
-  final String releaseDate;
-  final String producer;
-  final String plot;
-  final String url;
-  List<String> characters;
-  double? imdbRating;
-}
