+++ lib/ui/home/home_screen.dart
@@ -2,14 +2,14 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:star_wars_flutter/cache/shared_prefs.dart';
-import 'package:star_wars_flutter/presentation/model/movie_state.dart';
+import 'package:star_wars_flutter/ui/home/movies_state.dart';
 import 'package:star_wars_flutter/ui/common_widgets/empty_result_widget.dart';
 import 'package:star_wars_flutter/ui/common_widgets/errors_widget.dart';
 import 'package:star_wars_flutter/ui/common_widgets/loading_widget.dart';
 import 'package:star_wars_flutter/ui/common_widgets/platform_switch.dart';
 import 'package:star_wars_flutter/ui/home/movies_bloc.dart';
 import 'package:star_wars_flutter/ui/home/movies_widget.dart';
-import 'package:star_wars_flutter/ui/model/movie_view_model.dart';
+import 'package:star_wars_flutter/presentation/model/movie_view_model.dart';
 import 'package:star_wars_flutter/ui/theme/custom_theme.dart';
 
 class HomeScreen extends StatefulWidget {
