+++ lib/ui/home/movies_bloc.dart
@@ -2,8 +2,8 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:star_wars_flutter/domain/interactor/future_use_case.dart';
 import 'package:star_wars_flutter/domain/model/movie.dart';
 import 'package:star_wars_flutter/presentation/mapper/movie_mapper.dart';
-import 'package:star_wars_flutter/presentation/model/movie_state.dart';
-import 'package:star_wars_flutter/ui/model/movie_view_model.dart';
+import 'package:star_wars_flutter/ui/home/movies_state.dart';
+import 'package:star_wars_flutter/presentation/model/movie_view_model.dart';
 import 'package:star_wars_flutter/ui/utils/star_wars_image_utils.dart';
 
 import 'movies_event.dart';
