+++ lib/ui/movie_details/movie_details_bloc.dart
@@ -1,8 +1,8 @@
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:star_wars_flutter/domain/interactor/future_use_case.dart';
 import 'package:star_wars_flutter/domain/model/character.dart';
-import 'package:star_wars_flutter/presentation/model/movie_details_state.dart';
-import 'package:star_wars_flutter/ui/model/movie_view_model.dart';
+import 'package:star_wars_flutter/ui/movie_details/movie_details_state.dart';
+import 'package:star_wars_flutter/presentation/model/movie_view_model.dart';
 import 'package:star_wars_flutter/ui/utils/movie_utils.dart';
 
 import 'movie_details_event.dart';
