+++ lib/ui/movie_details/movie_details_widget.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:star_wars_flutter/ui/model/movie_view_model.dart';
+import 'package:star_wars_flutter/presentation/model/movie_view_model.dart';
 import 'package:star_wars_flutter/ui/movie_details/movie_details_content_widget.dart';
 
 class MovieDetailsWidget extends StatelessWidget {
